fix theme
