+++ next.config.js
@@ -1,11 +1,13 @@
-const { i18n } = require('./next-i18next.config.js')
+// const { i18n } = require('./next-i18next.config.js')
 
 const withPWA = require('next-pwa')({
   dest: 'public',
   register: true,
   skipWaiting: true,
 })
-
+// const withPWA = object => {
+//   return object
+// }
 // const withInit = require('next-in-it-stats')({ legacy: true })
 
 module.exports = withPWA({
