--- pages/_app.tsx
@@ -1,18 +1,0 @@
-import '../styles/variables.scss'
-import '../styles/globals.css'
-import type { AppProps } from 'next/app'
-import { SessionProvider } from 'next-auth/react'
-import { ThemeProvider } from 'hooks/useTheme'
-import { appWithTranslation } from 'next-i18next'
-import nextI18NextConfig from '../next-i18next.config.js'
-
-function MyApp({ Component, pageProps: { session, ...pageProps } }: AppProps) {
-  return (
-    <SessionProvider session={session}>
-      <ThemeProvider>
-        <Component {...pageProps} />
-      </ThemeProvider>
-    </SessionProvider>
-  )
-}
-export default appWithTranslation(MyApp, nextI18NextConfig)
